requirements.txt
pygame==1.9.6
is-wire==1.1.4
is-msgs==1.1.10
Pillow==5.3.0